pyproject.toml
[project]
name = "find-tunes"
version = "0.1.0"
description = "A shazam like app with machine learning based optimization for better audio recognization on noisey data"
readme = "README.md"
authors = [
    { name = "Akashthakur0326", email = "akashthakur0326@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "pandas>=2.2.3,<3.0.0",
    "yt-dlp",
    "nbstripout",
    "pydub",
    "python-dotenv>=1.2.1",
    "loguru",
    "librosa",
    "mlflow>=2.0.0",
    "onnxruntime",
    "protobuf<5.0.0",
    "setuptools>=82.0.0",
    "scipy>=1.17.0",
    "torchcrepe>=0.0.24",
    "torchaudio>=2.10.0",
    "torch>=2.10.0",
    "asyncpg",
    "sqlalchemy",
    "psycopg2-binary",
    "pgvector",
    "python-dotenv",
    "soundfile",
    "audiomentations",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]

[project.optional-dependencies]
backend = [
    "fastapi",
    "websockets",
    "pydantic",
    "python-multipart"
]
frontend = []

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
